# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a private API developed for Ajudaris",
        "version": "0.0.1",
        "title": "Ajudaris API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "127.0.0.1:3000",
...
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/admin/users/listAll": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Lists all users",
                "description": "Lists all users",
                "operationId": "listAllUsers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "type": "array"
                    },
                    "400": {
                        "description": "Error"
                    },
                    "401": {
                        "description": "Not Authorized"
                    }
                }
            }
        }
    }
}